Forsearching

+++ src/main/java/app/repository/VehicleRepository.java
@@ -1,9 +1,10 @@
 package app.repository;
 
 import app.libs.Vehicle;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VehicleRepository extends CrudRepository<Vehicle, Long> {
+public interface VehicleRepository extends CrudRepository<Vehicle, Integer> {
 }
